widgetise the messageList component
